% Generated by roxygen2 (4.0.1): do not edit by hand
\name{accession2gi}
\alias{accession2gi}
\title{Maps a genbank accession to a nuclotide database gi.}
\usage{
accession2gi(accession)
}
\arguments{
\item{accession}{accession character vector to lookup.}
}
\value{
named vector of gis.
}
\description{
Maps a genbank accession to a nuclotide database gi.
}

