% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{accession2taxid}
\alias{accession2taxid}
\title{Maps a nucleotide database accession to a taxonomy database taxId}
\usage{
accession2taxid(accessions)
}
\arguments{
\item{accessions}{accessions character vector to lookup.}
}
\value{
named vector of taxIds.
}
\description{
Maps a nucleotide database accession to a taxonomy database taxId
}
