% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{clustalo}
\alias{clustalo}
\title{Multiple sequence alignment with clustal omega}
\usage{
clustalo(x, exec = "clustalo", quiet = TRUE, original.ordering = TRUE,
  ...)
}
\arguments{
\item{x}{an object of class 'DNAbin'}

\item{exec}{a character string with the name or path to the program}

\item{quiet}{whether to supress output to stderr or stdout}

\item{original.ordering}{use the original ordering of the sequences}

\item{...}{additional arguments passed to the command line clustalo}
}
\description{
Calls clustal omega to align a set of sequences of class DNAbin.  Run
without any arguments to see all the options you can pass to the command
line clustal omega.
}

