% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterResults.R
\name{filter_PrimerTree_results}
\alias{filter_PrimerTree_results}
\title{Filter out sequences retrieved by search_primer_pair() that are either too
short or too long. The alignment and tree will be recalculated after removing
unwanted reads.}
\usage{
filter_PrimerTree_results(primerTreeObj, maxLength = 99999, minLength = 1)
}
\arguments{
\item{primerTreeObj}{a primerTree object}

\item{maxLength}{the maximum sequence length to keep}

\item{minLength}{the minimum sequence length to keep}
}
\value{
a primerTree object
}
\description{
Filter out sequences retrieved by search_primer_pair() that are either too
short or too long. The alignment and tree will be recalculated after removing
unwanted reads.
}

