% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterResults.R
\name{get_PrimerTree_Seq_Lengths}
\alias{get_PrimerTree_Seq_Lengths}
\title{Get sequence lengths from a primerTree object
Also returns a summary of the sequence lengths}
\usage{
get_PrimerTree_Seq_Lengths(primerTreeObj, summarize = TRUE)
}
\arguments{
\item{primerTreeObj}{a primerTree object}

\item{summarize}{a logical indicating if a summary should be displayed}
}
\value{
numeric values of sequence lengths
}
\description{
Get sequence lengths from a primerTree object
Also returns a summary of the sequence lengths
}

