% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{gi2taxid}
\alias{gi2taxid}
\title{Maps a nucleotide database gi to a taxonomy database taxId}
\usage{
gi2taxid(gi)
}
\arguments{
\item{gi}{gi character vector to lookup.}
}
\value{
named vector of taxIds.
}
\description{
Maps a nucleotide database gi to a taxonomy database taxId
}

