% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse_primer_hits}
\alias{parse_primer_hits}
\title{Parse the primer hits}
\usage{
parse_primer_hits(response)
}
\arguments{
\item{response}{a httr response object obtained from \code{\link{primer_search}}}
}
\description{
Parse the primer hits
}
