% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterResults.R
\name{seq_lengths.primerTree}
\alias{seq_lengths.primerTree}
\title{Method for primerTree objects}
\usage{
\method{seq_lengths}{primerTree}(x, summarize = TRUE)
}
\arguments{
\item{x}{a primerTree object.}

\item{summarize}{a logical indicating if a summary should be displayed}
}
\description{
Method for primerTree objects
}

