% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterResults.R
\name{table.primerTree}
\alias{table.primerTree}
\title{Get sequence lengths from a primerTree object
Also returns a summary of the sequence lengths}
\usage{
\method{table}{primerTree}(primertree_obj, summarize = TRUE)
}
\arguments{
\item{primertree_obj}{a primerTree object}

\item{summarize}{a logical indicating if a summary should be displayed}
}
\value{
sequence lengths
}
\description{
Get sequence lengths from a primerTree object
Also returns a summary of the sequence lengths
}
\examples{
# Show the counts for each length
table(mammals_16S)

# Plot the distribution of lengths
seqLengths <- table(mammals_16S)
barplot(seqLengths)
}

