% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterResults.R
\name{table}
\alias{table}
\title{Get sequence lengths from a primerTree object
Also returns a summary of the sequence lengths}
\usage{
table(...)
}
\arguments{
\item{primertree_obj}{a primerTree object}

\item{summarize}{a logical indicating if a summary should be displayed}
}
\description{
Get sequence lengths from a primerTree object
Also returns a summary of the sequence lengths
}

