% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{tree_from_alignment}
\alias{tree_from_alignment}
\title{Construct a neighbor joining tree from a dna alignment}
\usage{
tree_from_alignment(dna, pairwise.deletion = TRUE, ...)
}
\arguments{
\item{dna}{fasta dna object the tree is to be constructed from}

\item{pairwise.deletion}{a logical indicating if the distance matrix should 
be constructed using pairwise deletion}

\item{...}{furthur arguments to dist.dna}
}
\description{
Construct a neighbor joining tree from a dna alignment
}
\seealso{
\code{\link{dist.dna}}, \code{\link{nj}}
}
