% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterResults.R
\name{filter_seqs.primerTree}
\alias{filter_seqs.primerTree}
\title{Method for primerTree objects}
\usage{
\method{filter_seqs}{primerTree}(primertree_obj, min_length = 0,
  max_length = Inf)
}
\arguments{
\item{min_length}{the minimum sequence length to keep}

\item{max_length}{the maximum sequence length to keep}
}
\description{
Method for primerTree objects
}

