% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/filterResults.R
\name{filter_seqs}
\alias{filter_seqs}
\title{Filter out sequences retrieved by search_primer_pair() that are either too
short or too long. The alignment and tree will be recalculated after removing
unwanted reads.}
\usage{
filter_seqs(x, ...)
}
\arguments{
\item{x}{a primerTree object}

\item{min_length}{the minimum sequence length to keep}

\item{max_length}{the maximum sequence length to keep}
}
\value{
a primerTree object
}
\description{
Filter out sequences retrieved by search_primer_pair() that are either too
short or too long. The alignment and tree will be recalculated after removing
unwanted reads.
}
\examples{
# filter out sequences longer or shorter than desired:
mammals_16S_filtered <- filter_seqs(mammals_16S, min_length=131, max_length=156)
}

