% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxonomy.R
\name{get_taxonomy}
\alias{get_taxonomy}
\title{Retrieve the taxonomy information from NCBI for a set of nucleotide gis.}
\usage{
get_taxonomy(gis)
}
\arguments{
\item{gis}{a character vector of the gis to retrieve}
}
\value{
data.frame of the 'gis, taxIds, and taxonomy
}
\description{
Retrieve the taxonomy information from NCBI for a set of nucleotide gis.
}
