% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primerTree.R
\name{identify.primerTree_plot}
\alias{identify.primerTree_plot}
\title{identify the point closest to the mouse click
only works on single ranks}
\usage{
\method{identify}{primerTree_plot}(x, ...)
}
\arguments{
\item{x}{the plot to identify}

\item{...}{additional arguments passed to annotate}
}
\description{
identify the point closest to the mouse click
only works on single ranks
}

