% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{layout_tree_ape}
\alias{layout_tree_ape}
\title{layout a tree using ape, return an object to be plotted by
\code{\link{plot_tree}}}
\usage{
layout_tree_ape(tree, ...)
}
\arguments{
\item{tree}{The \code{\link{phylo}} tree to be plotted}

\item{...}{additional arguments to \code{\link{plot.phylo}}}
}
\value{
\item{edge}{list of x, y and xend, yend coordinates
as well as ids for the edges}
\item{tips}{list of x, y, label and id for the tips}
\item{nodes}{list of x, y and id for the nodes}
}
\description{
layout a tree using ape, return an object to be plotted by
\code{\link{plot_tree}}
}
